Rename file awsappsyncwebsockets.py to awsappsync.py
